Update to latest webext API
