--- lib/main.js
@@ -1,11 +1,0 @@
-var pageMod = require("sdk/page-mod");
-
-pageMod.PageMod({
-    include: "*.millionbook.com",
-    attachTo: ['top', 'existing'],
-    contentScriptFile: "./resizer.js",
-
-    onAttach: function(worker) {
-        worker.port.emit("init");
-    }
-});
