--- package.json
@@ -1,9 +1,0 @@
-{
-  "name": "millionbook-font-size",
-  "title": "millionbook-font-size",
-  "id": "jid1-OSbgFkKCB5twPA",
-  "description": "adds a font size changer on millionbook.com",
-  "author": "Benson Wong <mostlygeek@gmail.com>",
-  "license": "MPL 2.0",
-  "version": "0.3"
-}
